requirements.txt
# Core Django
Django==4.2.24
asgiref==3.9.1
sqlparse==0.5.3
typing_extensions==4.15.0

# Database
psycopg2-binary==2.9.10

# Password Hashing (match crystals app)
pbkdf2==1.3

# Django REST Framework & JWT Authentication
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1

# CORS Support
django-cors-headers==4.9.0

# Rate Limiting & Security
django-ratelimit==4.1.0

# Configuration Management
python-decouple==3.8

# HTTP Requests (for testing scripts)
requests==2.32.5
certifi==2025.10.5
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0